routes: migrate skills + workflows domains (10 routes)
Move to:
- src/server/routes/skills.ts (2): /api/slash-skills, /api/slash-skills/details
- src/server/routes/workflows.ts (7): /api/workflows CRUD + customize/override

Drop now-unused inner closures (resolveProjectConfigStore, resolveSkillDiscoveryCwd)
and now-unused imports (discoverSlashSkills, getSkillDirectories) from server.ts.

Net: server.ts -171 lines (6594 → 6423). No behaviour change.
