+++ src/server/routes/dispatcher.ts
@@ -23,6 +23,8 @@ import { modelsRoutes } from "./models.js";
 import { preferencesConfigRoutes } from "./preferences-config.js";
 import { rolesRoutes } from "./roles.js";
 import { toolsRoutes } from "./tools.js";
+import { skillsRoutes } from "./skills.js";
+import { workflowsRoutes } from "./workflows.js";
 
 // Registered route arrays — populated as domain migrations land. Each
 // `routes/<domain>.ts` exports a `<domain>Routes: Route[]` array. Until all
@@ -37,6 +39,8 @@ const allRoutes: ReadonlyArray<Route> = [
 	...preferencesConfigRoutes,
 	...rolesRoutes,
 	...toolsRoutes,
+	...skillsRoutes,
+	...workflowsRoutes,
 ];
 
 /**
